Split SQL to submodules
