+++ src/daf/logging/sql/__init__.py
@@ -1,0 +1,7 @@
+"""
+SQL logging package.
+"""
+
+from .mgr import *
+if SQL_INSTALLED:
+    from .tables import *
